+++ src/components/preslang_old/op/Init_global_var.js
